+++ frontend/src/lib/components/Navigation.svelte
@@ -6,7 +6,7 @@
 	<nav class="list-nav p-4">
 		<ul>
 			<li><a href="/client/upload-csv-test">Upload CSV Test</a></li>
-			<li><a href="/client/transactions/budget/1">Table Test</a></li>
+			<li><a href="/client/budget/1">Table Test</a></li>
 			<li><a href="/client/accounts">Accounts</a></li>
 			<li><a href="/client/budget/mine">Budget</a></li>
 			<li><a href="/client/skelly">Skelly</a></li>
